added another index to make easier searching in scene titles and stage directions
You can change in which index to search by changing the index String
field in the search classes.

"index_simple" is the indexing for all the contents of the file and it
is used in simple search tab and when we are searching for all fields in
advanced search

"index" is the indexing that has all speech tags indexed and it is used
in advanced search when searching in dialogues

"index_scenes" is the indexing that indexes all scenes and holds fields
for scene titles and stage directions. This is used when searching in
scene titles and stage directions in advanced search